// src/app/opmobilitybackend/models/get-consumable-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ConsumableImageResponse } from '../models/consumable-image-response';
import { GetExpoEventResponse } from '../models/get-expo-event-response';
import { ProductResponse } from '../models/product-response';
export interface GetConsumableResponse {
  category: string;
  event: GetExpoEventResponse;
  eventId: number;
  id: number;
  images: Array<ConsumableImageResponse>;
  name: string;
  notes: string;
  product: ProductResponse;
  productId: number;
  providerId: number;
  purchasedAt: string;
  quantity: number;
  unit: string;
  unitCost: number;
}